+++ core/src/backend_impl/macroquad/window.rs
@@ -2,9 +2,9 @@ use crate::math::Size;
 pub use macroquad::miniquad::conf::Icon as WindowIcon;
 use macroquad::window::{screen_height, screen_width};
 
-pub fn window_size() -> Size<u32> {
+pub fn window_size() -> Size<f32> {
     Size {
-        width: screen_width() as u32,
-        height: screen_height() as u32,
+        width: screen_width(),
+        height: screen_height(),
     }
 }
